add converter to handle timespans as seconds
the response json has timestamps as numbers - we need to convert those
to timespans using a custom converter

found it impossible to run the tests, as they require connecting to services
that require authentication, and I don't have the credentials. this must
be done even for small, separate tests - so i had to create a new test-project
that can run without any external dependencies.